Moves tbc-debug-util source and header files to /src and /include
