+++ SConstruct
@@ -82,7 +82,7 @@ src_files = Split("""src/TEM.cpp
                      src/snowsoil/Stefan.cpp
                      src/snowsoil/TemperatureUpdator.cpp
                      src/CrankNicholson.cpp
-                     src/util/tbc-debug-util.cpp
+                     src/tbc-debug-util.cpp
                      src/vegetation/Vegetation_Bgc.cpp
                      src/vegetation/Vegetation_Env.cpp""")
